Force old behavior of binutils/linker
Our issue is similar to this [1] issue. Though we shouldn't have
undefined weak symbols, but I think it is somehow connected. This should
be revisited when a new version of binutils comes out, as the issue
should be solved by the new version [2]. The solution is coming from
the official recommendtation [3].

[1] https://github.com/msys2/MINGW-packages/issues/7023
[2] https://github.com/msys2/MINGW-packages/pull/8259
[3] https://www.msys2.org/news/#2021-01-31-aslr-enabled-by-default